[Fix] ProblemAnalysisService self-invocation 트랜잭션 분리 수정
- ProblemAnalysisDbService 신규 빈으로 fetchAnalysisPreparation, saveAnalysisSuccess 분리
  → analyzeProblemSync에서 직접 self-call 대신 별도 빈 호출로 AOP 프록시 정상 적용
- analyzeProblemSync에 @Transactional(NOT_SUPPORTED) 추가
  → OpenAI 호출 중 DB 커넥션 미점유 설계 의도 실제 동작
- saveAnalysisSuccess에 COMPLETED 상태 중복 체크 추가
  → 동시 처리 시 이미 완료된 분석 덮어쓰기 방지
- PROBLEM_ANALYSIS_NOT_FOUND 예외를 Consumer와 Service 양쪽에서 graceful skip 처리
  → 문제 삭제 후 Phase 3 실행 시 불필요한 DLQ 처리 방지
